change folder structure for readability
