+++ .gitignore
@@ -6,3 +6,6 @@ mainClusterer/mainClusterer
 test_client/test_client
 clusteringNode/clusteringNode
 output.log
+client 
+introducer
+mainClusterer
