+++ go.mod
@@ -1,15 +1,12 @@
-module ridehost
+module github.com/sujitv19196/ridehost
 
-go 1.19
-
-require gonum.org/v1/gonum v0.12.0
+go 1.20
 
 require (
-	github.com/biogo/cluster v0.0.0-20150317015930-38d63f016191 // indirect
-	github.com/google/uuid v1.3.0 // indirect
-	github.com/muesli/clusters v0.0.0-20200529215643-2700303c1762 // indirect
-	github.com/muesli/kmeans v0.3.1 // indirect
-	github.com/nextmv-io/sdk v0.22.0 // indirect
-	github.com/nu7hatch/gouuid v0.0.0-20131221200532-179d4d0c4d8d // indirect
-	golang.org/x/exp v0.0.0-20230321023759-10a507213a29 // indirect
+	github.com/google/uuid v1.3.0
+	github.com/muesli/clusters v0.0.0-20200529215643-2700303c1762
+	github.com/muesli/kmeans v0.3.1
+	gonum.org/v1/gonum v0.13.0
 )
+
+require golang.org/x/exp v0.0.0-20230321023759-10a507213a29 // indirect
